fixed/some issues

+++ app/Http/Controllers/BooksController.php
@@ -6,6 +6,8 @@ use App\Http\Library\ApiHelpers;
 use App\Models\Book;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Pagination\Paginator;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
@@ -16,98 +18,15 @@ class BooksController extends Controller
 {
     use ApiHelpers;
 
-
-
-    /////////////////// HERE /////////////////////
-
-//
-//
-//    public function update(Request $request){
-//        $post = Post::find($request->id);
-//        // check if user is editing his own post
-//        // we need to check user id with post user id
-//        if(Auth::user()->id != $post->user_id){
-//            return response()->json([
-//                'success' => false,
-//                'message' => 'unauthorized access'
-//            ]);
-//        }
-//        $post->desc = $request->desc;
-//        $post->update();
-//        return response()->json([
-//            'success' => true,
-//            'message' => 'post edited'
-//        ]);
-//    }
-//
-//    public function delete(Request $request){
-//        $post = Post::find($request->id);
-//        // check if user is editing his own post
-//        if(Auth::user()->id !=$post->user_id){
-//            return response()->json([
-//                'success' => false,
-//                'message' => 'unauthorized access'
-//            ]);
-//        }
-//
-//        //check if post has photo to delete
-//        if($post->photo != ''){
-//            Storage::delete('public/posts/'.$post->photo);
-//        }
-//        $post->delete();
-//        return response()->json([
-//            'success' => true,
-//            'message' => 'post deleted'
-//        ]);
-//    }
-//
-//    public function posts(){
-    //    $posts = Post::orderBy('id','desc')->get();
-    //    foreach($posts as $post){
-    //        //get user of post
-    //        $post->user;
-    //        //comments count
-    //        $post['commentsCount'] = count($post->comments);
-    //        //likes count
-    //        $post['likesCount'] = count($post->likes);
-    //        //check if users liked his own post
-    //        $post['selfLike'] = false;
-    //        foreach($post->likes as $like){
-    //            if($like->user_id == Auth::user()->id){
-    //                $post['selfLike'] = true;
-    //            }
-    //        }
-
-    //    }
-
-    //    return response()->json([
-    //        'success' => true,
-    //        'posts' => $posts
-    //    ]);
-//    }
-
-    ////////////////// HERE /////////////////////
-
-
-
-
-
-
-
-    public function book(Request $request): \Illuminate\Http\JsonResponse
+ public function book(Request $request): \Illuminate\Http\JsonResponse
     {
 
-            // $post = DB::table('books')->get();
-            // return $this->onSuccess($post, 'Book Retrieved');
-
-
-            $perPage = $request->input('page', 10);
+            $perPage = $request->input('page', 5);
             if (!is_numeric($perPage)) {
-              $perPage = 10;
+              $perPage = 5;
             }
 
-            // $books = Book::withCount('likes')->orderBy('likes_count', 'desc')->paginate($perPage);
-            $books = Book::withCount('likes')->orderBy('likes_count', 'desc')->get();
+            $books = Book::withCount('likes')->orderBy('likes_count', 'desc')->paginate($perPage,2,1);
             foreach($books as $book){
                 //get user of bo
                 $book->user;
@@ -115,28 +34,34 @@ class BooksController extends Controller
                 $book['commentsCount'] = count($book->comments);
                 //likes count
                 $book['likesCount'] = count($book->likes);
-             
+
                 //check if users liked his own post
                 $book['selfLike'] = false;
                 foreach($book->likes as $like){
                 if($like->user_id == Auth::user()->id){
                     $book['selfLike'] = true;
                 }
-            }                
             }
-
-        
+        }
             return response()->json([
                 'status' => 200,
                 'message' => "Retrieved",
                 'data' => $books
             ]);
 
-
-
     }
 
+    public function paginate($items,$perPage,$pageStart=1)
+    {
 
+        // Start displaying items from this number;
+        $offSet = ($pageStart * $perPage) - $perPage;
+
+        // Get only the items you need using array_slice
+        $itemsForCurrentPage = array_slice($items, $offSet, $perPage, true);
+        return new LengthAwarePaginator($itemsForCurrentPage, count($items),
+            perPage,Paginator::resolveCurrentPage(), array('path' => Paginator::resolveCurrentPath()));
+    }
 
     public function createBook(Request $request): \Illuminate\Http\JsonResponse
     {
@@ -145,7 +70,8 @@ class BooksController extends Controller
        if ($this->isAdmin($user)){
            $validator = Validator::make($request->all(), $this->postValidationRules());
            if ($validator->passes()){
-               // Create New Book
+               if(Auth::user()->role==1){
+                // Create New Book
                $book = new Book();
                $book->user_id = Auth::user()->id;
                $book->title = $request->input('title');
@@ -155,6 +81,10 @@ class BooksController extends Controller
                $book->user;
 
                return $this->onSuccess($book, 'Book Created');
+               
+               }else{
+                return $this->onError(400, 'Only Admin Access needed');
+               }
            }
            return $this->onError(400, $validator->errors());
        }
@@ -169,13 +99,17 @@ class BooksController extends Controller
         if ($this->isAdmin($user)){
             $validator = Validator::make($request->all(), $this->postValidationRules());
             if ($validator->passes()){
-                // Create New Book
+            if(Auth::user()->role==1){
+                // update  Book
                 $book = Book::find($request->id);
                 $book->title = $request->input('title');
                 $book->content = $request->input('content');
                 $book->update();
 
                 return $this->onSuccess($book, 'Book Updated');
+            }else{
+                return $this->onError(400, 'Only Admin Access needed');
+               }
             }
             return $this->onError(400, $validator->errors());
         }
@@ -185,16 +119,20 @@ class BooksController extends Controller
 
     public function deleteBook(Request $request): \Illuminate\Http\JsonResponse
     {
-        
+
         $user = $request->user();
-        // if ($this->isAdmin($user) || $this->isUser($user)){
+        if ($this->isAdmin($user) ){
+            if(Auth::user()->role==1){
             $book = Book::find($request->id); // Find the id of the post passed
             $book->delete();         // Delete the specific post data
             if (!empty($book)){
                 return $this->onSuccess($book, 'Book Deleted');
             }
             return $this->onError(404, 'Book Not Found');
-        // }
+           }else{
+            return $this->onError(400, 'Only Admin Access needed');
+           }
+        }
         return $this->onError(401, 'Unauthorized Access');
     }
 
@@ -204,6 +142,7 @@ class BooksController extends Controller
         if ($this->isAdmin($user)){
             $validator = Validator::make($request->all(), $this->userValidatedRules());
             if ($validator->passes()){
+                if(Auth::user()->role==1){
                 // Create New Writer
                 User::create([
                     'name' => $request->input('name'),
@@ -214,6 +153,9 @@ class BooksController extends Controller
 
                 $writerToken =  $user->createToken('auth_token', ['writer'])->plainTextToken;
                 return $this->onSuccess($writerToken, 'User Created With Writer Privilege');
+               }else{
+                return $this->onError(400, 'Only Admin Access needed');
+               }
             }
             return $this->onError(400, $validator->errors());
         }
@@ -222,29 +164,5 @@ class BooksController extends Controller
 
     }
 
-    public function createSubscriber(Request $request): \Illuminate\Http\JsonResponse
-    {
-        $user = $request->user();
-        if ($this->isAdmin($user)){
-            $validator = Validator::make($request->all(), $this->userValidatedRules());
-            if ($validator->passes()){
-                // Create New Writer
-                User::create([
-                    'name' => $request->input('name'),
-                    'email' => $request->input('email'),
-                    'role' => 3,
-                    'password' => Hash::make($request->input('password'))
-                ]);
 
-                $writerToken =  $user->createToken('auth_token', ['subscriber'])->plainTextToken;
-                return $this->onSuccess($writerToken, 'User Created With Subscriber Privilege');
-            }
-            return $this->onError(400, $validator->errors());
-        }
-
-        return $this->onError(401, 'Unauthorized Access');
-
-    }
-
-  
 }
